Add files via upload
This C++ program defines a class car that represents a car with three properties: brand, model, and year. It includes a member function displayinfo() that outputs the car's details. The program then creates three different car objects (car1, car2, and car3), each representing a specific car with unique brand, model, and year attributes. The details of each car are displayed using the displayinfo() function. The cars used in this example are a ford Model S (2020), mercedes AM4 (2018), and a Lamborghini  svj 63(2022).